Use high score list names for Astral player mons
Use a random name from the high score list for player monsters on
Astral, if possible.  If no appropriate name can be found (probably
because the topten/record file is empty), just use the standard
developer names as before.  There is no consideration for or attempt to
match the role or sex (or race, obviously) of the hero in the topten
entry, out of concern this might make it too difficult to find a
matching entry for a particular player monster.

No explicit Is_astral(&u.uz) check is necessary since get_mplname is
only called for player monsters generated on Astral (unlike in, say,
EvilHack, which does something similar with naming Astral player mons).