Get unique topten names a more efficient way
Still try to ensure the high score list names for player monsters are
only used once each, to avoid saturating a level with duplicates of one
player who plays/ascends a lot, but do it a bit more efficiently than
just picking a random name and re-rolling if it isn't unique.  Instead,
select a topten entry and then use the first unique/unused name between
that entry and the sysopt max topten rank.

The function is basically get_rnd_toptenentry, but with the additional
"keep reading the next entries until you get something unique, if
requested to do so" bit.