Swapchest: return to using item num to pick prefix
This behavior (the number of items you've put into the swapchest
determining the name, with names getting more grandiose as the number
increases -- e.g. the first item inserted in a particular game is named
"a token from Foo", then the second is named "generously donated by
Foo", etc) was ditched in 788ce63, which changed the swapchest item file
to store the player name alone, without any prefix, so the name could be
checked and compared to the current player's.  Add the donation count to
the swapchest file format, so that the old behavior of fixed names based
on how many others were donated can be restored.

There was also an issue with using '%ms' -- it is a unique feature of
glibc, so didn't work on my laptop running OSX, and I think it leaked
memory because the variable 'donorname' was never free'd after being
allocated by sscanf (for those using glibc, at least).  Since the player
name is fixed/limited to PL_NSIZ anyway, just use a buffer of fixed size
for 'donorname'.