Make vanilla achievements formatting match website
Bring the formatting of the names and descriptions of vanilla
achievements in-game into sync with how they are represented on the
website, since updating the game is easier than updating the site
mid-tournament (and I think a number of these changes are improvements
anyway, though that is just my opinion).  This should be revisited
after the tournament, and maybe some mix of the site and game versions
should be settled on.