Fix: "crushed to death by a crushed"
If the player kills a DevTeam member, they are permanently killed by a
dungeon collapse.  To ensure this isn't survivable, there are two done()
calls in a row, so a hero with lifesaving will get killed by the second
one.  However, done() clears killer.name, so a hero who survived the
initial collapse of the dungeon would end up with the death reason
"crushed by a crushed", thanks to the default killer.name from done().