Game over if DevTeam member killed by monster
The "rocks fall, you die" ending for killing a DevTeam member applied
only if the player directly did the deed (it was in xkilled), so
training up a purple worm to a high level and letting 'er rip provided a
way to kill off Pat Rankin without any punishment (besides the
admittedly severe intrinsic punishment of losing Pat).  Collapse the
dungeon if a DevTeam member is killed, whether by the player or by a
monster.