Fix: impossible when saving after arena cowardice
Fix "impossible bad monster weapon restore" when a player saves her game
after fleeing an arena fight (it was due to stale mtmp->mw pointer when
it referred to a transient item that had been deleted).  This should fix
it (and should be fine to leave in place),  but I feel like maybe the
entire level should be deleted when the player leaves the arena, so it
isn't included in savefiles at all beyond that point.  Not sure.