Adjust feedback when mon eats swap chest
I just found this on a branch from last year.  Don't leave monster's
disappearance unremarked on if the player has 'verbose' off, use a
different message if you can see the spot but the monster is invisible
so you can't see the action, and break one of the messages into two so
it has a better chance of fitting into 80 characters.