Fix discoveries list achivement progress ordering
For the multiple achievements in gem class, once you had discovered
enough gems the achievement progress lines were being interspersed with
the discoveries list in a way that seemed random.  They were ordered
based on otyp rather than position in the discoveries list, so it wasn't
actually random, but the difference in the two sort orders made it seem
that way (and made them hard to find in the list).  Print them all at
the end of the class instead.