Fix: Mikado achievement was broken
Both scrolls and wands have extra items listed in objects.c that are
just placeholders for extra appearances to be shuffled in. Scrolls
accounted for this when counting them up for the achievement. Wands
didn't.

Because the impetus here is a couple players who have got all 24 wands
and who would now be unable to earn the achievement, I added a temporary
shim that will allow them to earn it by doing something that would
result in makeknown(some wand otyp).