Restore missing nullptr guard to previous commit
montraits() may return (struct monst *) 0, so dereferencing its return
value without first making sure it's non-null was dangerous and, well,
just not the hottest idea overall.

This moves the addition of the _BY_HERO bit further down, which kills
two birds with one stone.  It's past an early return if mon is null,
which serves as the guard against null pointer dereference that was
needed.  On top of that, the 'by hero' bit can now be set even if
used_saved_traits was FALSE for the statue reanimation.  Or that's true
in theory, at least -- as of right now, I don't think mon->mrevived will
ever wind up being set if execution didn't proceed via the montraits()
call.