Adjust permitted swap chest items
Make some adjustments to the items that can be stashed in the swapchest,
based in part on observations of and by players from last year:

 * Reject +0 or worse charged rings
 * Allow scaring mundane instruments (bugle, horn, drum -- I think
   people will actually find these useful and it won't just be an item
   dump but can be reconsidered)
 * Allow non-dull athames

Another request was to allow touchstones, which could help people ID
gems and put together protection money, but I think that would encourage
archeologists to dump their touchstone into the swapchest and produce an
overwhelming amount of touchstones.  It would also leak some information
about gray stone identities, I guess.  So I'm not doing that right now.

I also reformatted swap_chest_eligible to be more consistent with the
code formatting/style used in the rest of the codebase.