Weight swap chest destinations towards source
Instead of an inserted swap chest item going to one of the three servers
purely at random, weight the destinations so that it will most likely
(60%) go to the current server (i.e. the one the player is playing on
when inserting the item).  This has the advantage of automatically
weighting destinations based on server population, hopefully mitigating
the issue last year where tons of good swapchest items piled up in AU
since not enough people were playing to take advantage of them, while
the most populous servers were constantly scraping the bottom of the
barrel^Wswapchest: more populated servers will have more players
inserting items, and most of those items will stay on the server where
they were inserted.

This is a bit of an experiment and the actual weights may need to be
adjusted, but I think it's a good idea overall and am interested to see
what the results will be.