Include "scuffed" artifact onames in NPC file
"Grayswandir" will become "Graysvandi?" or something (similar to when
your "hand slips" if attempting to use an artifact name); that way you
can recognize which artifact(s) the hero you're facing actually ascended
with.

There is maybe some potential for confusion: someone could not be
reading too closely, and get a particularly similar-looking result from
wipeout_text (like "Vcrpal Blade") and think they'll be able to snag an
artifact from the pile, giving up other items that they would actually
prefer if they hadn't been misled.  We can revert it if that starts
happening with any frequency, but hopefully it will be fun instead of
frustrating.