Adjust "Prize Inside" award criteria, description
Don't give player the achievement for breaking a statue of a monster she
stoned herself with footrice corpse/eggs.  Also tweak the achievement
description so it reflects that the statue does not need to be
"randomly-generated" (i.e. not guaranteed/specified in the des file for
the level), but only "naturally-generated".  That term is not really
great but I'm not sure how else to put it without the description
becoming very wordy.