Adjust award criteria for "Captain Ahab"
Award A_SURVIVED_DROWNING whenever you get unstuck from a sea monster;
the previous method (becoming unstuck whenever the monster was on a
water square) left some loopholes that should probably have awarded the
achievement but didn't: for example, freezing the water beneath an eel
that's grabbing you, and killing it on the resulting ice.  The
achievement description doesn't require imminent drowning anyway, just
that you were grabbed by a sea monster, so I don't think it's a problem
that this change no longer has a guarantee built in that the hero
escaped an imminent drowning (and I don't think an eel or kraken will
grab you outside of the water anyway, so that guarantee may still exist
in practice anyway).

In addition to the "freezing the water" escape method, another one that
didn't grant the achievement was level teleportation away from the grab.
This should grant it under both scenarios (and hopefully others that
haven't been reported or identified yet, if they exist).