Don't set cknown for empty swap chest
Under other circumstances the swapchest cknown bit would not be set, but
if the chest was empty it would be.  Don't set it when empty either.
Also, even if cknown is set somehow, don't hide the "take items out"
options in the #loot menu if the swap chest was empty.  Even if the
chest was empty when the menu was put together, a new item could have
been added by the time the player selects an option.