Fix: 'both' loot mode vs swap chest
Selecting 'b' for the swap chest allowed players to put an item into the
swap chest after taking an item out.  This would increment swapitems to
0 (from -1, SWAP_CHEST_USED_UP) and reset the swap chest to its original
state, so could be repeated multiple times to keep the swap chest
working (though you'd end up paying at least two items per each swap
item taken -- one for the initial reset and a second one to actually get
it to open again).  On my local copy the game actually crashed due to an
asan problem this casued, but I think on the server this would work and
allow indefinite use of the swapchest.  This should prevent it from
happening.