Shorten swap chest item removal failure messages
Several of them exceeded 80 characters, sometimes by quite a lot, so
would wrap to multiple lines if playing with the minimum recommended
terminal size of 80x26.  Also, rework the failure messages so the array
contains only a shorter reason for failure, not an entire sentence: that
way they can be plugged into a sentence that actually refers to the item
by name (or typename), instead of using "the item" or "the object"
everywhere.

I tried to keep the reasons somewhat close to the old messages, at least
in spirit, but with the significant reduction in length a lot of
zaniness and character was lost -- I think that's just the price of
sticking to the normal length for in-game messages, though.