Revise "Horatius at the Bridge"
Kops can't ever become extinct in normal play, so update the achievement
to reflect this.  While we're at it, rephrase the description a bit
more, since extinct isn't used as a verb in standard English, and use
the same phrasing for "Dudley's Hero".  I didn't touch the use of
"extinct" in the achievement enum IDs even though a vision of
"A_EXTIRPATED_foo" danced before my eyes to tempt me for a second.