Block "Minesweeper" if monster blew up Ludios mine
The "fairness" exception that kept "Minesweeper" achievable if a monster
set off a land mine in Ludios was causing some inconsistencies.  If a
monster blew up the very last mine on the level, the achievement was
impossible to get, but if the monster blew up the second-to-last mine
(or every mine on th level but one) and you disarmed the last one, you'd
still get it.  This makes the achievement harder, but I think it
shouldn't be impossible to achieve in most Ludios games if you're quick
about it.