Disallow levelporting from ongoing deathmatch
Not entirely happy about this but I found that branchporting is already
forbidden, so we'll stick with that for now.  The problem with allowing
escape from other sources than the portal is that it's not feasible to
seal off the arena level when it happens, unless the destination happens
to be the level with the arena portal.  I have been thinking that maybe
we could force the destination of levelport or branchport to be the
arena portal level and then move the portal-sealing and item-erasure
parts from domagicportal to goto_level, but this is a simple initial fix
at least.