Make swapchest inedible
This is not a very "fun" approach to fixing this issue, but we can
always replace it later with some funny message if it seems wise.  The
swapchest is already inedible when it has contents, so just extend that
inedibility to the swap chest at all times.  This prevents griefing by
eating the valley swapchest and dying, thereby depriving the recipient
of your bones of the valley chest.