Restrict #tnntdebug access during beta period
Add a TNNTDEVS sysconf entry and match the username against that to
restrict #tnntdebug access only to users in that list.  The command is
not meant for public access, and the only reason it is enabled outside
of wizard mode when TNNT_BETA is defined is to allow devs to do testing
directly on the server.