Notify user of completed achievement
This will print a message when the user has completed an achievement, if
she turns on the associated option ("tnnt_notify", off by default).
This doesn't work for the "vanilla achievements" (the first group listed
on the website and in #achievements, #V01-V11), because of how those are
tracked in-game.  That may be a possible extension in the future, but I
think it would probably involve redoing how vanilla vs TNNT achievements
are tracked.  The phrasing of "TNNT achievement unlocked: %s" is a
departure from normal messages in game, but I think we can live with it.
It's an unusual kind of meta-game message anyway.