Add ubirthday to swapobj filenames
plname + o_id isn't a guarantee of uniqueness since objects can be
donated across multiple games, meaning a later donation could
potentially overwrite an earlier one.  Cross-server donations mean that
theoretically this isn't any guarantee of uniqueness either, since two
games could be started on multiple servers by the same person at the
same second, but the likelihood of that causing a problem is remote and
the consequences would be mild, so I'm not worrying about it right now.