Move (again) and fix "Peak Condition" check
Placing it in bot() was causing the tnnt_notify pline to appear in the
wrong place -- over the map, or in the statuslines -- because it was
interrupting some statusline or map refresh type action.  Move the test
for "Peak Condition" to a once-per-action check in moveloop instead,
which should hopefully work better.

While moving it, I realized it was iterating through A_MAX (A_CHA+1) to
A_STR+1, instead of A_CHA to A_STR.  I think this means A_STR was not
being checked, and some value one past the end of u.acurr.a[] (perhaps
u.aexe.a[0]) was checked instead.  This should fix that too.