Prevent nymphs from picking up the swap chest
I thought this was not an issue until 3.7, because non-tame nymphs don't
pick up containers, but Furey noticed that tame nymphs will still pick
it up.  This is a problem because you can get your pet nymph to move the
swap chest to another (bones-eligible) dungeon level, die, and the
recipient of those bones will have 3 swap chests in their game.  Make
the swap chest impossible for monsters to carry, even nymphs.