Automatically insert year, beta status in dumplog
The year in the dumplog title was being manually updated, and was easy
to forget about.  Insert the year that the game was started instead.
This isn't ideal because it means someone playing the version from the
previous year's tournament would have the wrong year in their dumplog
(e.g. it would say "The November NetHack Tournament 2024" if they were
playing the version that was live during 2023), but since the variant is
designed for server play where that won't happen, I think having one
less thing to forget about when the next year rolls around is probably
worth it.  Maybe we can think of another way to do it later that avoids
that issue without requiring manual updates each year.

Also distinguish between games played during the beta and the actual
tournament.