Treat NKIs like real objects in #terrain filter
When the player selects "hide monsters and objects" in #terrain, hide
all the non-kitten items (and the kitten) on the robotfindskitten level.
I didn't think this would be very useful since it just makes the level
completely empty, but apparently someone had a difficult time finding
the portal back to the Dungeons of Doom because #terrain wasn't helping
to filter out all the NKI clutter.