Restrict "Diamond in the Rough" grant conditions
Only grant it if actually identifying a previously-unidentified gem --
require the stone you're IDing not already be identified, and also
restrict it only to gemstones (so don't grant it for using a touchstone
on a flint stone).  The second part is arguable since it's not clear
whether "gem" in the description refers to a member of the gem class or
an actual gem, but since glass was already excluded and the title
mentions diamonds I'm taking it to mean an actual gemstone.