More swap chest item removal logging
Handle the explore mode (or wizard mode, but that's not very relevant)
situation of the player removing her own item from the swap chest, and
provide a more complete description of the item, including enchantment,
fireproofing, etc.  This ended up being a lot more code in the end than
I would have thought; if the added detail in the messages doesn't feel
necessary we can definitely revisit and cut it down considerably.