Refine conditions for awarding Most Dangerous Game
It was being awarded for reaching the point where a monster stepped on a
trap you set, even though the monster may have been saved by a
subsequent roll (or was too large for the trap, etc) so not actually
trapped.  Even though this is a lot less clean, I think it has to be
split up like this into the various specific traps, since many of them
don't actually set mtrapped.  [Then again, it could be argued that it
should only apply to traps that do set mtrapped, since "being caught in
a trap" sounds more to me like being held in the trap than
e.g. triggering some kind of explosive.]

I think this should cover everything but it's possible a bug or two
snuck in because of the increase in complexity.