Permit autopicking transient items during fight
The reason these were blocked from autopickup is to spare the hero from
stepping onto the pile, automatically picking up a scroll, and losing
the chance to actually select an item.  Go ahead and allow autopickup of
transient items (probably ammo or missiles that the Arena NPC has fired
at the hero) while the fight is in progress, since they will be handled
appropriately after the fight ends even if they're in the player's
inventory, and won't do any harm to the loot selection process after a
deathmatch victory.