Tweak arena victory "pile" messaging a bit more
Give the message about something disappearing from your inventory even
if that's the only transient item on the map (unlikely...).  If there's
only one item and you can see it appear on the ground, still print a
message about it, and print an appropriate message that doesn't
reference "choosing" a prize from the pile of items.