Notify of completed (tracked) vanilla achievements
Somewhat common feedback we received last year was that the "achievement
unlocked" message ought to show the completion of vanilla achievements,
too.  Some players expected it and thought the achievement wasn't
tracked when it didn't show up.

We had discussed trying to integrate vanilla achivements into the TNNT
achievement system, to address that and avoid needing to manually sync
the vanilla achievements lists in the backend and game.  That's still an
option, but the current achievement numbering, which distinguishes
vanilla achivements by numbering them separately and with a 'V' prefix,
complicates it.

This is a quicker, simpler alternative approach that prints the message
for vanilla achievements without big changes to the achievement system.