Implement two additional achievements
These were sitting in my local tnnt copy since at least last year's
tournament - one to find a Vlad trapdoor without triggering it, and one
to hit a monster with a diagonal ray that has bounced 3 or more times.

Also tweak the description of "Pop the Question" to clarify that
seduction is part of it - otherwise you might think you ought to get it
by simply throwing a ring of adornment to a succubus (who will ignore
it).