Close off a loophole with "never killed X" conducts
Since the conducts only checked mvitals[].died, it was possible to turn
some of these monsters into something else, then kill it. This did not
count as killing a member of that species, so the conduct remained
intact.

Solve this by tracking the monster ID when a unique relevant to one of
the conducts is created, and flag it when a monster with that ID is
killed. (It strikes me as incorrect to increment mvitals[].died, since
the kill would then count for two separate species. Also, for
achievements which require you to kill a certain amount of species, you
arguably didn't kill the species that is supposed to count.)