Hide certain lines in #tnntstats until they become available
There are several things tracked by tnntstats that are not useful to
show from the start of the game because the hero can't yet directly make
any progress towards them, and so they only serve to clutter the list.

I hid the ones that have a clear and easily-tracked point at which they
can start being worked on. These are:
 - Mysterious force levels won't show up until you have the Amulet.
 - High altar visits won't show up until you start the ascension run.
 - Number of times you killed the Wizard, ditto.
 - Ludios landmines won't show up until you enter Ludios.
 - Planes elementals won't show up until you're in the Planes.
 - Purple worms on Astral won't show up until Astral.

I debated also hiding things like the wish sources until you first get a
wish, and so on, but that logic would extend to most of the other things
shown in #tnntstats, making most of its contents hidden by default.

In the end-of-game summary and dumplog, no lines will be hidden.