Restructure the two passtune-based achievements
qt pointed out that these achievements are nearly redundant, since
almost everyone who uses the passtune learns it by deducing it and then
immediately gets the achievement for opening the drawbridge with it.

I modified it so that they are now entirely distinct achievements, to
the point you can't get them both in the same game: learning the
passtune from your god will either take some work (or will be a
consolation prize if you are pray-farming for spellbooks) and the other
one is changed to *deducing* the passtune, which remains a worthy
achievement. Merely using the passtune to open the bridge would follow
from either of those so I dropped it as an achievement.