Fix: "Get Flashed" was awarded when hallucination obscured flash
When you are hallucinating and drop a blessed or cursed item on an
altar, it flashes a random color, and you know that it is not uncursed,
but have no further information and the game doesn't set bknown. But the
achievement was being awarded anyway.

Bring it into line with the achievement description (learn the beatitude
from an altar, even if that beatitude is uncursed) by testing if the
beatitude changed from 0 to 1.

Note that prior to this commit, you could also get the achievement by
dropping an already BUC-identified item on an altar. Now you can't; you
actually have to learn its beatitude.