Add 6 more achievements
Drawing on achievement suggestions that gained support: get enough globs
to make a large glob, tame a Rider, and identify various sets of tools.

That last one (4 different achievements for bags, horns, light sources,
and magic music instruments) hadn't been implemented along with other
"identify a set of objects" because it wasn't clear what the best way to
do it was. All non-obvious tools would have virtually required
polypiling or wishing for items to complete the set, so I peeled off
bags and light sources. Horns were a bit tricky since they can be
instruments, but all the magic ones have a different use, so I ended up
peeling those off as well. Left with only whistles and tonal
instruments, there were a few options: 1) have a separate achievement
for identifying both of each set of two items, 2) have a single
achievement for identifying all eight of them, or 3) have an achievement
for identifying the four magic ones, which is what I went with. I'm not
convinced it was the right choice because you don't have to identify the
nonmagic instruments, but option 1 felt like bloating the achievement
list and option 2 felt like it would still involve polypiling, wishing,
or extinctionist tactics to get all the requisite instruments.

This pushes the achievements into the tnntachieve5 range as now there
are more than 320 of them. Not expected to require any special work done
on the scoreboard.

Also noticed a couple things with existing identification achievements
that will break in 3.7, so stuck TODOs in there for it.