Fix: Mr. Bones' Wild Ride could be awarded with < 4 bones piles
... via amnesia. The counter tracking how many bones piles you found did
not change when you forgot a bones pile.

Fix this by adding a decrement when a bones pile is forgotten. As qt
pointed out, a "proper" fix for this would probably involve tracking a
struct cemetery linked list, but for now this is all it should need, and
as described in the comment, this fix will go away when 3.7 is merged
when amnesia goes away.