Fix: Izchak's Apprentice should outright ignore brass lantern
I mistakenly thought having the achievement defined by a range that
happened to include the always-identified brass lantern wouldn't be a
problem. It's not a problem in terms of earning the achivement, but it
creates a display problem: since the discoveries list command shows
progress towards completion when any items that count towards it are
identified, it would show "Izchak's Apprentice: 1/5" from the start of
the game and apparently attached to the wrong object class unless some
tools started out identified.

This change makes Izchak's Apprentice display only after some lamp or
candle is identified and as n/4 instead of n/5, which is accurate in
terms of the number of things you have to identify but not accurate in
terms of the achievement definition of "identify all light sources". But
it wouldn't improve anything to change that definition to a much wordier
"identify all light sources which don't start identified"; oh well.