Fix: achievement progress didn't display properly in discovery list
Bug report was that upon getting two discoveries that counted towards
"The One-Man Magic Band", the discoveries list showed two progresses
towards it, both with 1/2 instead of a single entry with 2/4.

I realized this is because the code that tracks achievement progress
for display purposes in dodiscoveries() was set up to handle a single
pass through the list of objects, assuming that all items counting
towards an achievement appear in a contiguous blob - allowing for some
gaps with items that count towards no achievement, but not interwoven
with items that count towards other achievements. But the items required
for The One-Man Magic Band ARE scattered around another achievement; the
magic whistle and flute come before the horns (Four Horns Adventures)
and the magic harp and drum of earthquake come after them. This was no
longer possible to do in a single pass without some method of storing
achievements that had already been visited in the past.

I went with a generalized solution rather than a hack for this one
achievement - within each object class, up to 4 achievements can be
tracked and the order of objects no longer matters. Then the summaries
for those achievements are printed at the end of the object class, same
as before, except this time storing whole strings isn't needed.