Do not generate dumplog if game was startscummed.
If the player quits or escapes within 100 turns or less, the dumplog
files (.txt and .html) are not generated. This is done to combat serial
startscummers from generating huge amounts of junk dumplog files. Saves
volume space and is a bit less load on the server.

This currently prevents dump_open_log() from executing, and guards
against dump_close_log(). Potential for improvement here possibly.

Co-authored-by: entrez <me@entrez.cc>