Migrate old TNNT design discussions into top level .md file
I was trying to move in other design discussions besides ones about
achievements, but quickly ran into the problem that there were no
obvious, easily findable places to put those other topics. Burying
deathmatch discussions in comments around the deathmatch code, for
instance, is the opposite of discoverable. So instead, I created a
readable, visible file to hold all of them (including the achievements
ones, which I moved out of tnnt_achievements.h).

A benefit of having this file is the ability to give it a preamble and
establish some basic things about it, like the time range the
discussions happened in. The bit about summarizing instead of copying is
important, I think - not only do I want to make this more readable, I
don't want to take the text of private comments and make them public.