Fix: No Rulez being awarded in bones Minetown empty of watchmen
This has been a known bug for at least a couple years. At first it was
much more prominent because it also occurred in Orcish Town and we did a
quick and dirty patch that ignored that level variant for the
achievement, but the case of getting a bones Minetown someone already
cleared still existed.

This solves the problem and removes the dirty patch by checking to see
if there are no watchmen in Minetown when it's first loaded. If so, it
sets a new flag that blocks the achievement from being awarded.

As noted in a comment, if we want to, we could allow the player to earn
the achievement by resurrecting or revgenoing a watchman to clear that
flag, then killing or moving them out of Minetown, but that feels rather
forced and I didn't feel like implementing it at this time.