Adjust criteria for Boy Scout achievement
A suggestion from ais523 - Boy Scout is intended to be for finishing all
the Sokoban puzzles without cheating, but was actually implemented as
finishing the *branch* without cheating, since it triggered upon picking
up the prize. In addition, you could obtain it without actually solving
the lowest level, since that is passable via walking through the pits.
This commit changes it so that it gets awarded as soon as the game
recognizes all four levels as solved (and the player still hasn't broken
any rules).

Meanwhile, the Boulder Pusher achievement remains unchanged, still
triggering upon obtaining the Sokoban prize. That one is for completing
the branch, which does involve making it through the zoo.

There's an edge case oddity in how the new Boy Scout behavior works:
because it's based in the mapseen logic, if the hero completes some but
not all Sokoban levels, then gets mind flayed and forgets some of them,
then goes on to complete the final levels without ever revisiting the
forgotten ones, the achievement won't be awarded until they revisit all
four levels and the mapseen list marks them all as solved at the same
time. While it would have been possible to design around this (i.e.
saving the ledger number of each level in tnnt_globals once it's
completed or something similar), I think the chance of it happening is
sufficiently low that we don't need to worry about it. Even if you did
get mind flayed at the wrong time, you would have to levelport past
forgotten levels to avoid scoring the achievement. Also, once 3.7 is
released and TNNT gets based on it, it will be impossible to forget
levels anyway.