Adjust a few achievement names and descriptions
Suggestion by ais523: make the description for Touch It With A 10 Foot
Pole more explicit about what its requirements are.

Suggestion by qt: revert Glass Menagerie back to Glass Collector since
the former name breaks the pattern of other identify-all-X achievements,
which are written like titles that could be given to the hero.

Suggestion by ais523: make explicit that the resurrect the Wizard
achievement means resurrect his corpse, and won't be earned if you
petrify and unpetrify him.

Also, write down a new achievement idea in the once-we-are-on-3.7
section.