Move most TNNT functions into new tnnt.c file
As described at the top of tnnt.c, this code move is so that we can keep
TNNT code (or at least TNNT functions) separate from the rest of the
game code and reduce the amount of merge conflicts down the line (as
functions move around and split up in vanilla). There will still be
plenty of merge conflicts for intra-function TNNT code, but we can't do
much about that.

A few functions had to remain outside the file because they relied on
specific things only available in their source file and which it didn't
seem wise to move into headers.