Add README.md as a TNNT specific readme
This is added as the TNNT specific readme because I noticed if you visit
the page on github it still just renders the 2023 nethack 3.6.7 README
(which I did not touch).

The basic template is from the xNetHack README but of course its content
is heavily modified. I made a point of mentioning that we only really
support linux with text windowports but do maintain the default tileset
for webtiles.