Make Mike Stephenson's throne give no effects
This throne being basically a free extra throne undefended by hostile
monsters made it one of the biggest ways the game balance of TNNT was
skewed relative to regular NetHack, and wishes or genocides from it were
common. To correct this while leaving the throne in the game, I've now
special-cased it to print a message when sat on.

Also removed the possibility of a hero kicking the throne for goodies,
and confused looting the throne to generate monsters. This throne
shouldn't do anything.

Since there are now no effects, you also don't earn the achievement for
sitting on a throne with it (otherwise you could earn the achievement
with no risk).