Track vanilla achievements as regular TNNT achievements
This resolves a five year old TODO. Previously, vanilla achievements
displayed their number as "#V__"; now they are just achievements #001
through #011.

The main reason for doing this is to simplify several key pieces of code
both here in the game and on the scoreboard side. There's not a
compelling enough reason to treat them differently and maintain separate
arrays (separate yaml files and xlogfile processing on the scoreboard)
and glue code so they get treated the same as other achievements, when
we could just... make them ordinary TNNT achievements.