Don't print "Achievement unlocked" for Mines' End luckstone
There are a couple Mines' End designs that try to trick the player into
walking away with a loadstone or flint stone. Getting the "achievement
unlocked" message gave it all away that you had found the luckstone, and
made Mines' End a bit easier compared to vanilla.

As noted in the comment, there are ways around this, but this makes it
at least a little better.