Capitalize all words in achievement names and add style guide
Achievement names were inconsistent in whether or not they capitalized
words (typically stop words). This establishes a standard that all words
should be capitalized in an achievement name, unless it makes it
nonsensical when capitalized. (The one case of this - "Team a
Headquarters" - I changed the name to "Team Ant Headquarters" rather
than leaving it as a weird-looking exception. I could see us possibly
turning "Learn The Alphabet", which is about killing lowercase monsters,
into "learn the alphabet" at some future point, as well.)

Other changes in this commit:
- Deleted a comment talking about putting vanilla achievements in here
  (this was done in eebc63d)
- Moved the list of TNNT achievements superseded by 3.7 ones to the
  bottom, and updated it by removing the ones which are completely
  duplicated by TNNT achievements now with identical criteria. Kept the
  ones that have different criteria, which will prompt a decision on
  whether to change them to be the same as 3.7's or keep the TNNT ones.
- Added a impossible-in-3.6 achievement idea to the 3.7 list.